Former-commit-id: cdb4ae52be8ed229a9d2948b95e5e335dd7743d3
Former-commit-id: 09289048b3bf518054f511eb3a9cdfdcb5174b94

+++ src/ui/colorAnimator.h
@@ -13,10 +13,10 @@ namespace phi
     {
     private:
         static std::unordered_map<color*, colorAnimation*> _animations;
-        static int _lastUpdateMilliseconds;
+        static double _lastUpdateMilliseconds;
 
     public:
-        static void animateColor(color* colorFrom, color colorTo, int milliseconds);
+        static void animateColor(color* colorFrom, color colorTo, double milliseconds);
         static void animateColor(colorAnimation* animation);
         static void update();
     };
